refactor: Define the limit-preservation map once
Before, we were defining the limit-preservation map ad-hoc for each
functor in question.  This is dubious, as these maps have the correct
type, but a priori they are not uniquely determined.

Now, we define `pres F : F (Lim F) → ShLim F` once for a general functor
F, and reuse it for all other proofs.  This is mostly a straightforward
search-and-replace, except for `Bag`.  There, we construct an explicit
fixpoint and show that the underlying map is pres.  Before, this was
true definitionally, but now only holds up to a 2-path `p ≡ refl ∙ p`.

In any case, the definitions are now more consistent and easier to use.